refactoring to pull the new storage handling to pyramids as well

+++ include/tdp/features/brief.h
@@ -46,9 +46,9 @@ namespace tdp {
     int minId = -1;
     int minDist = 257;
     if (a.IsValid()) {
-      for (int i=0; i<bs.w_; ++i) {
+      for (size_t i=0; i<bs.w_; ++i) {
         // iterate over pyramid levels
-        for (int j=0; j<bs.h_; ++j) 
+        for (size_t j=0; j<bs.h_; ++j) 
           if (bs(i,j).IsValid()) {
             int dist = Distance(a.desc_, bs(i,j).desc_);
             if (dist < minDist) {
@@ -66,7 +66,7 @@ namespace tdp {
     int minId = -1;
     int minDist = 257;
     if (a.IsValid()) {
-      for (int i=0; i<bs.size(); ++i) 
+      for (size_t i=0; i<bs.size(); ++i) 
         if (bs[i]->IsValid()) {
           int dist = Distance(a.desc_, bs[i]->desc_);
           if (dist < minDist) {
